+++ test/selenium.test.js
@@ -1,31 +1,39 @@
-import assert from "node:assert";
 import path from "node:path";
 
-import { By } from "selenium-webdriver";
+import selenium from "selenium-webdriver";
 import chrome from "selenium-webdriver/chrome.js";
-import { describe, it } from "vitest";
+import { afterAll, beforeAll, describe, expect, it } from "vitest";
 
 import util from "../src/util.js";
 
-const { Driver, ServiceBuilder, Options } = chrome;
-
-describe("run", async () => {
+describe("run", async function () {
   let driver = undefined;
 
-  it("should run post install", async () => {
-    const options = new Options();
-    const args = [
+  beforeAll(async function () {
+    const options = new chrome.Options();
+    const seleniumArgs = [
       `--nwapp=${path.resolve("test", "app")}`,
       "--headless=new",
     ];
-    options.addArguments(args);
 
-    const chromedriverPath = util.findpath("chromedriver");
+    options.addArguments(seleniumArgs);
 
-    const service = new ServiceBuilder(chromedriverPath).build();
+    const chromedriverPath = await util.findpath("chromedriver", { flavor: 'sdk' });
+    const service = new chrome.ServiceBuilder(chromedriverPath).build();
 
-    driver = Driver.createSession(options, service);
-    const text = await driver.findElement(By.id("test")).getText();
-    assert.strictEqual(text, "Hello, World! Is anyone out there?");
+    driver = chrome.Driver.createSession(options, service);
+  });
+
+  it("should run post install", async function () {
+    const textElement = await driver.findElement(selenium.By.id('test'));
+
+    const text = await textElement.getText();
+
+    expect(text).toEqual('Hello, World! Is anyone out there?');
   }, { timeout: Infinity });
+
+  afterAll(function () {
+    driver.quit();
+  });
+
 });